+++ Sources/Internal/Math/Matrix2.h
@@ -1,7 +1,7 @@
 #ifndef __DAVAENGINE_MATH2DMATRIX2_H__
 #define __DAVAENGINE_MATH2DMATRIX2_H__
 
-#include <math.h>
+#include <cmath>
 #include "Base/BaseTypes.h"
 #include "Math/MathDefines.h"
 #include "Math/MathConstants.h"
@@ -92,8 +92,8 @@ inline void Matrix2::SetIdentity()
 
 inline void Matrix2::BuildRotation(float32 angle)
 {
-    float32 cosA = cosf(angle);
-    float32 sinA = sinf(angle);
+    float32 cosA = std::cos(angle);
+    float32 sinA = std::sin(angle);
 
     data[0] = cosA;
     data[1] = sinA;
