+++ Sources/Internal/Render/2D/Systems/VirtualCoordinatesSystem.h
@@ -135,12 +135,12 @@ inline const Rect& VirtualCoordinatesSystem::GetFullScreenVirtualRect() const
 
 inline float32 VirtualCoordinatesSystem::AlignVirtualToPhysicalX(float32 value) const
 {
-    return floorf(value / physicalToVirtual + 0.5f) * physicalToVirtual;
+    return std::floor(value / physicalToVirtual + 0.5f) * physicalToVirtual;
 }
 
 inline float32 VirtualCoordinatesSystem::AlignVirtualToPhysicalY(float32 value) const
 {
-    return floorf(value / physicalToVirtual + 0.5f) * physicalToVirtual;
+    return std::floor(value / physicalToVirtual + 0.5f) * physicalToVirtual;
 }
 
 inline float32 VirtualCoordinatesSystem::ConvertPhysicalToVirtualX(float32 value) const
